requirements.txt
asgiref==3.5.2
boto3==1.18.23
botocore==1.21.65
Brotli==1.0.9
CacheControl==0.12.11
cachetools==4.2.4
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
date-time-literal==1.0.7
Django==3.0.7
django-compression-middleware==0.4.1
django-cors-headers==3.7.0
django-crontab==0.7.1
django-model-utils==4.2.0
django-notifications-hq==1.7.0
django-notifications-rest==0.0.1
django-storages==1.11.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.7.2
drf-yasg==1.20.0
facebook-sdk==3.1.0
fcm-django==0.3.11
firebase-admin==5.2.0
future==0.18.2
google-api-core==2.8.2
google-api-python-client==2.51.0
google-auth==2.8.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.1
google-cloud-firestore==2.5.3
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.3
grpcio==1.46.3
grpcio-status==1.46.3
httplib2==0.20.4
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==0.10.0
jsonfield==3.1.0
MarkupSafe==2.1.1
msgpack==1.0.4
mysqlclient==2.1.0
oauthlib==3.2.0
packaging==21.3
Pillow==9.1.1
proto-plus==1.20.6
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyfcm==1.4.9
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-twitter==3.5
pytz==2021.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.5.2
six==1.15.0
sqlparse==0.4.2
swapper==1.1.2.post1
uritemplate==4.1.1
urllib3==1.26.9
zstandard==0.17.0
